requirements.txt
flask
flask-cors
langchain==0.1.20
chromadb==0.4.24
python-dotenv
gdown
google-generativeai
langchain-google-genai
langchain-community
sentence-transformers
